bot/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
iso8601==2.1.0
multidict==6.0.5
python-dateutil==2.9.0.post0
six==1.16.0
twitchio==2.8.2
typing_extensions==4.10.0
yarl==1.9.4
